+++ Packages/com.espale.utilities/CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2025-09-18
+
+### Added
+
+- Added `MathUtilities.ClampAngleNormalized` method. It works the same way as `MathUtilities.ClampAngle` but it normalizes the angle beforehand
+- Added `RandomUtilities.Shuffle` and `RandomUtilities.GetShuffled` methods. They both take in `List`s or `Array`s as parameters and shuffle them. `RandomUtilities.Shuffle` replaces the original collection while `RandomUtilities.GetShuffled` returns a new shuffled version of the original collection.
+
 ## [1.0.6] - 2025-01-10
 
 ### Added
 
-- Added `MathUtilities.IsInClosedRange`. Does the same thing as `MathUtilities.IsInRange` but the range is closed (end points are included in the range.).
+- Added `MathUtilities.IsInClosedRange` method. Does the same thing as `MathUtilities.IsInRange` but the range is closed (end points are included in the range.).
 
 ## [1.0.5] - 2024-12-25